Loop and Termination Key
I added a loop to the main part of the program and the ability to terminate the program at the press of the key q. This should satisfy the ability to start and stop at the same place once I add the ability to deque the entire queue into a text file and reque on starting it up again.